@@ -1,12 +1,16 @@
+#!/usr/bin/env python3
 # -*- coding: utf-8 -*-
-import os
+from pathlib import Path
+from typing import Optional
 import pandas as pd
 from dateutil import parser
 # ruff: noqa: RUF100
-from getfactormodels.models.models import (barillas_shanken_factors,  # noqa: F401, E501
-                                           carhart_factors, dhs_factors,
-                                           ff_factors, hml_devil_factors,
-                                           icr_factors, liquidity_factors,
+from getfactormodels.models.models import \
+    barillas_shanken_factors  # noqa: F401
+from getfactormodels.models.models import carhart_factors  # noqa: F401, E501
+from getfactormodels.models.models import (dhs_factors, ff_factors,
+                                           hml_devil_factors, icr_factors,
+                                           liquidity_factors,
                                            mispricing_factors,
                                            q_classic_factors, q_factors)
 from getfactormodels.utils.cli import parse_args
@@ -14,10 +18,10 @@ from getfactormodels.utils.utils import _get_model_key, _process
 
 
 def get_factors(model: str = "3",
-                frequency: str = "M",
-                start_date=None,
-                end_date=None,
-                output=None) -> pd.DataFrame:
+                frequency: Optional[str] = "M",
+                start_date: Optional[str] = None,
+                end_date: Optional[str] = None,
+                output: Optional[str] = None) -> pd.DataFrame:
     """Get data for a specified factor model.
 
     Return a DataFrame containing the data for the specified model and
@@ -58,6 +62,7 @@ def get_factors(model: str = "3",
         raise ValueError(f"Invalid model: {model}")
 
     df = function(frequency, start_date, end_date, output)
+
     return df
 
 
@@ -78,11 +83,11 @@ class FactorExtractor:
     """
 
     def __init__(self,
-                 model='3',
-                 frequency='M',
-                 start_date=None,
-                 end_date=None,
-                 output=None):
+                 model: str = '3',
+                 frequency: Optional[str] = 'M',
+                 start_date: Optional[str] = None,
+                 end_date: Optional[str] = None,
+                 output: Optional[str] = None):
         self.model: str = model
         self.frequency: str = frequency
         self.start_date = self.validate_date_format(start_date) if start_date \
@@ -91,14 +96,19 @@ class FactorExtractor:
             else None
         self.output = output
         self._no_rf = False
+        self._no_mkt = False
         self.df = None
 
-    def no_rf(self):
+    def no_rf(self) -> None:
         """Sets the _no_rf flag to True."""
         self._no_rf = True
 
+    def no_mkt(self) -> None:
+        """Sets the _no_mkt flag to True."""
+        self._no_mkt = True
+
     @staticmethod
-    def validate_date_format(date_string):
+    def validate_date_format(date_string: str) -> str:
         """
         Validate the date format.
 
@@ -108,7 +118,8 @@ class FactorExtractor:
         try:
             return parser.parse(date_string).strftime("%Y-%m-%d")
         except ValueError as err:
-            raise ValueError("Incorrect date format, use YYYY-MM-DD.") from err
+            error_message = "Incorrect date format, use YYYY-MM-DD."
+            raise ValueError(error_message) from err
 
     def get_factors(self) -> pd.DataFrame:
         """Fetch the factor data and store it in the class."""
@@ -116,15 +127,22 @@ class FactorExtractor:
             model=self.model,
             frequency=self.frequency,
             start_date=self.start_date,
-            end_date=self.end_date)
+            end_date=self.end_date,
+            output=self.output)
 
         if self._no_rf:
-            self.df = self.drop_rf(self.df)
+            self.df = self.drop_rf(self.df.copy())  # create a copy before drop
+        if self._no_mkt:
+            self.df = self.drop_mkt(self.df.copy())
 
         return self.df
 
-    def drop_rf(self, df):
+    def drop_rf(self, df: pd.DataFrame = None) -> pd.DataFrame:
         """Drop the ``RF`` column from the DataFrame."""
+        # get_factors if not already done
+        if df is None:
+            df = self.get_factors()
+
         if "RF" in df.columns:
             df = df.drop(columns=["RF"])
         else:
@@ -132,7 +150,19 @@ class FactorExtractor:
 
         return df
 
-    def to_file(self, filename):
+    def drop_mkt(self, df: pd.DataFrame = None) -> pd.DataFrame:
+        """Drop the ``MKT`` column from the DataFrame."""
+        if df is None:
+            df = self.get_factors()
+
+        if "Mkt-RF" in df.columns:
+            df = df.drop(columns=["Mkt-RF"])
+        else:
+            print("`drop_mkt` was called but no MKT column was found.")
+
+        return df
+
+    def to_file(self, filename: str):
         """
         Save the factor data to a file.
 
@@ -153,12 +183,15 @@ def main():
                                 start_date=args.start, end_date=args.end)
     if args.no_rf:
         extractor.no_rf()
+    if args.no_mkt:
+        extractor.no_mkt()
 
     df = extractor.get_factors()
 
     if args.output:
         extractor.to_file(args.output)
-        print(f'File saved to "{os.path.abspath(args.output)}"')
+        print(f'File saved to "{Path(args.output).resolve()}"')
+
     else:
         print(df)
 
