@@ -10,8 +10,8 @@
 # copies of the Software, and to permit persons to whom the Software is
 # furnished to do so, subject to the following conditions:
 #
-# The above copyright notice and this permission notice shall be included in all
-# copies or substantial portions of the Software.
+# The above copyright notice and this permission notice shall be included in
+# all copies or substantial portions of the Software.
 #
 # THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
 # IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
@@ -20,10 +20,10 @@
 # LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 # SOFTWARE.
-__version__ = "0.0.3"
+__version__ = "0.0.4"
 
 from .__main__ import FactorExtractor, get_factors
-from .models import models  # noqa: F401
+from .models import models  # noqa: F401, RUF100 (silent flake8 in VScode)
 from .models.models import (barillas_shanken_factors, carhart_factors,
                             dhs_factors, ff_factors, hml_devil_factors,
                             icr_factors, liquidity_factors, mispricing_factors,
