@@ -2,7 +2,7 @@
 import argparse
 
 
-def parse_args():
+def parse_args() -> argparse.Namespace:
     """Argument parser, allowing for command line arguments.
     This is the function used in pyproject.toml to run the CLI."""
     parser = argparse.ArgumentParser(
@@ -21,8 +21,10 @@ def parse_args():
                         help='The start date for the data.')
     parser.add_argument('-e', '--end', type=str, required=False,
                         help='The end date for the data.')
-    parser.add_argument('-o', '--output', type=str, required=False,  # noqa
+    parser.add_argument('-o', '--output', type=str, required=False,
                         help='The file to save the data to.')
     parser.add_argument('--no_rf', '--no-rf', '--norf', action='store_true',
                         help='Drop the RF column from the DataFrame.')
+    parser.add_argument('--no_mkt', '--no-mkt', '--nomkt', action='store_true',
+                        help='Drop the Mkt-RF column from the DataFrame.')
     return parser.parse_args()
